Windows port: ttnn_core first-wave MSVC fixes
- operation.hpp: include <experimental/type_traits> only where it exists
- serialization / overlapped_serialization: mmap-based tensor loads read the
  file into memory on Windows (no mmap); POSIX fd/stat path kept elsewhere
- tensor_impl / h2d_socket_service: POSIX includes behind _WIN32 guards;
  page size via GetSystemInfo
- graph_processor: std::array iterators are not raw pointers on MSVC's STL
- explicit narrowing casts in braced initializers (reports, graph_trace_utils,
  tensor_spec_flatbuffer, tensor_utils, xtensor/partition) -- MSVC hard-errors
  where GNU/Clang builds run with -Wno-narrowing

(cherry picked from commit 10aa73922491c7fe8b6974c8678abfe812fcd3c4)

+++ ttnn/core/tensor/flatbuffer/tensor_spec_flatbuffer.cpp
@@ -15,7 +15,9 @@ CoreRangeSet from_flatbuffer(const flatbuffer::CoreRangeSet* core_range_set) {
     ranges.reserve(core_range_set->ranges()->size());
     for (const auto* range : *core_range_set->ranges()) {
         ranges.emplace_back(
-            tt::tt_metal::CoreCoord{range->start()->x(), range->start()->y()}, tt::tt_metal::CoreCoord{range->end()->x(), range->end()->y()});
+            tt::tt_metal::CoreCoord{
+                static_cast<size_t>(range->start()->x()), static_cast<size_t>(range->start()->y())},
+            tt::tt_metal::CoreCoord{static_cast<size_t>(range->end()->x()), static_cast<size_t>(range->end()->y())});
     }
     return CoreRangeSet{std::move(ranges)};
 }
